refactor: orders앱 order로 수정
- domain도 orders -> order로 수정
- mysql내부의 장고관련 테이블 속성들도 수정
