+++ order/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from orders.views import CartView, CartsView
+from order.views import CartView, CartsView
 
 urlpatterns = [
     path('/carts', CartsView.as_view()),
