+++ order/views.py
@@ -4,7 +4,7 @@ from django.http                import JsonResponse
 from django.views               import View
 from django.db.models           import Sum
 
-from orders.models         import Cart
+from order.models         import Cart
 from core.token_decorators import token_decorator
 
 class CartView(View):
