+++ product/views.py
@@ -6,7 +6,7 @@ from django.db.models           import Count, Q, Sum, Case, When, F, Prefetch
 from django.db.models.functions import Coalesce
 
 from product.models import Category, SubCategory, Product, Item, ProductComponent
-from orders.models   import Cart
+from order.models   import Cart
 
 class CategoryView(View):
     def get(self, request):
